week1/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.1
openai==1.68.2
python-dotenv==1.1.0
